Reorganise tests
Also Sublime highlighting :b:roke, but

Ref: #9
Ref: #40
