+++ tests/base/from_bytes.rs
