+++ tests/base/from_bytes_pedantic.rs
