+++ tests/base/transmute_many.rs
