+++ tests/base/transmute_many_pedantic.rs
