+++ tests/bool/mod.rs
@@ -1,0 +1,4 @@
+mod vec_permissive;
+mod vec_pedantic;
+mod permissive;
+mod pedantic;
