+++ tests/bool/pedantic.rs
