+++ tests/bool/vec_pedantic.rs
