+++ tests/bool/vec_permissive.rs
