+++ tests/full/many_pedantic.rs
