+++ tests/full/many_permissive.rs
