+++ tests/full/mod.rs
@@ -1,0 +1,6 @@
+mod many_permissive;
+mod many_pedantic;
+mod one_pedantic;
+mod many;
+mod one;
+mod vec;
