+++ tests/full/one_pedantic.rs
