+++ tests/full/vec.rs
